fix toggling to also add class to sections
added TODO note - we should reuse same code in MFE to take advantage
of its unit tests. We also want to be able to deal with toggling and
dynamically loading content in the mobile site so if there is a problem
in toggle.js for MFE that prevents this we should get that fixed.